+++ app/Http/Controllers/AssetController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\AssetRequest;
 use App\Http\Resources\AssetCollection;
 use App\Models\Asset;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 
@@ -12,6 +13,8 @@ class AssetController extends Controller
 {
     public function index()
     {
+        $this->authorize('showAsset',User::class);
+
         $requestParam = \request()->all('search', 'trashed');
         return Inertia::render('Asset/Index', [
             'filters' => $requestParam,
@@ -26,11 +29,15 @@ class AssetController extends Controller
 
     public function create()
     {
+        $this->authorize('createAsset',User::class);
+
         return Inertia::render('Asset/Create');
     }
 
     public function store(AssetRequest $request)
     {
+        $this->authorize('createAsset',User::class);
+
         Asset::create(
             $request->validated()
         );
@@ -46,6 +53,8 @@ class AssetController extends Controller
 
     public function edit(Asset $asset)
     {
+        $this->authorize('editAsset',User::class);
+
         return Inertia::render('Asset/Edit',[
             'asset' => $asset
         ]);
@@ -53,6 +62,8 @@ class AssetController extends Controller
 
     public function update(AssetRequest $request, Asset $asset)
     {
+        $this->authorize('editAsset',User::class);
+
         $asset->update(
             $request->validated()
         );
@@ -62,6 +73,8 @@ class AssetController extends Controller
 
     public function destroy(Asset $asset)
     {
+        $this->authorize('deleteAsset',User::class);
+
         $asset->delete();
 
         return to_route('asset.index');
