+++ app/Http/Controllers/RegisterTokenController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Resources\RegisterTokenCollection;
 use App\Models\RegisterToken;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 
@@ -11,6 +12,8 @@ class RegisterTokenController extends Controller
 {
     public function index()
     {
+        $this->authorize('showToken', User::class);
+
         return inertia('RegisterToken', [
             'tokens' => new RegisterTokenCollection(
                 RegisterToken::query()
@@ -20,22 +23,28 @@ class RegisterTokenController extends Controller
         ]);
     }
 
-    public function create(){
+    public function create()
+    {
+        $this->authorize('createNotice', User::class);
+
         $uuid = Str::uuid();
-        if(!RegisterToken::query()->where('uuid',$uuid)->exists()){
+        if (!RegisterToken::query()->where('uuid', $uuid)->exists()) {
             RegisterToken::query()->create([
                 'uuid' => $uuid,
                 'expire_at' => now()->addDay()
             ]);
-        }else{
+        } else {
             self::create();
         }
-        return back()->with('success','New Token Created');
+        return back()->with('success', 'New Token Created');
     }
 
-    public function destroy(Request $request){
-        $request->validate(['id'=>'required']);
+    public function destroy(Request $request)
+    {
+        $this->authorize('deleteNotice', User::class);
+
+        $request->validate(['id' => 'required']);
         RegisterToken::query()->whereId($request->id)->delete();
-        return back()->with('success','Token Delete');
+        return back()->with('success', 'Token Delete');
     }
 }
