+++ app/Http/Controllers/RuleController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\RuleRequest;
 use App\Http\Resources\MessCollection;
 use App\Http\Resources\RuleCollection;
 use App\Models\Rule;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 
@@ -13,6 +14,8 @@ class RuleController extends Controller
 {
     public function index()
     {
+        $this->authorize('showRule',User::class);
+
         $requestParam = \request()->all('search', 'trashed');
         return Inertia::render('Rule/Index', [
             'filters' => $requestParam,
@@ -29,11 +32,15 @@ class RuleController extends Controller
 
     public function create()
     {
+        $this->authorize('createRule',User::class);
+
         return Inertia::render('Rule/Create');
     }
 
     public function store(RuleRequest $request)
     {
+        $this->authorize('createRule',User::class);
+
         Rule::create(
             $request->validated()
         );
@@ -49,6 +56,8 @@ class RuleController extends Controller
 
     public function edit(Rule $rule)
     {
+        $this->authorize('editRule',User::class);
+
         return Inertia::render('Rule/Edit', [
             'rule' => $rule,
         ]);
@@ -56,6 +65,8 @@ class RuleController extends Controller
 
     public function update(RuleRequest $request, Rule $rule)
     {
+        $this->authorize('editRule',User::class);
+
         $rule->update(
             $request->validated()
         );
@@ -65,6 +76,8 @@ class RuleController extends Controller
 
     public function destroy(Rule $rule)
     {
+        $this->authorize('deleteRule',User::class);
+
         $rule->delete();
 
         return to_route('rule.index');
